Boss Summoned based on number killed
NOTES: I changes the rate at which the player fires.
currently the spawn is set to 5. we can always change it.
We should add more bosses and vary the shot speeds.
Also currently once the enemy destroyed reaches the apex it spawns the boss and whatever is on screen stays.