Fixed code a bit
Managed to fixed the issue of tracking with all the ships. Still need to fix the issue of the ships not being able to be controlled after spawning in again. Or the information showing up again after having spawned in.. We can't add the highscore feature until we fix this. Will work more on the code tommorow