added highscore and powerups
Note that there are issues with both of the codes at the moment. The powerups with constantly spawning in everytime i hit the space button and the highscore with crashing the game. Though that's more so how I implimented it...It's still writing to a text file which is good for now.