firing type and cleanup?
Got the triple shot working and cleaned the whole firing code up. you do not need to check each one only 2 and then the last case is a default case.
Currently the first case is to check for ship 0 because it is faster to load in that one. We definitely need to change that out later.
Furthermore I will be working on dual shot forward and making audio additions and  comment them out for now. Will add tag that will signify what lines need to be uncommented for later. Also will start working on pattern and  firing for another boss  and how they look when damaged. REMINDER SCALE BACK DAMAGE ON SEPERATE BULLETS.