highscore edits
The highscore is now being read to a file, however it still needs to be shown in game so that is something I have to work on

on top of that issue occurs with three lives being lost as if we were to try and remove the player...the enemy ships will not be able to find the player and as such cause an index out of bounds error. This could probably be fixed if we remove the enemies as well. Or even just move the player off the screen to another spot.

Added a loose message to test some things, note that this is only a test object and can be removed at any time during the testing