Loss message properly functions
The appropriate order of if statements had to be in place. Then after removing the player the enemies needed to all die or else they would crash due to nothing being on screen. Lastly needed to make sure the enemies stopped spawning. also will need to do this for the boss. Unfortunately I was at work and had more important work to attend to. Please apply similar code for the boss to either stop shooting or be killed.