changes made to code
ShipInfoScreen although still there is now the ship selection screen with some added text. will have to work more on that later and figure out how to fix the spacing issues. Removed up to x10 powerup and left it up to x4 as it seems to cause major overflow...the highscore also updates on the mainscreen now as it turns out it was just reording of code with the file reader having to be up on the top.