Fixed some Screens around
Fixed the UI screens around, tell me what you think in a text message or in a document. Removed the ShipInfoScreen as we already have that set as the shipselection screen and no longer need two seperate things. Will have to continue to work on edits with the powerup and others later.