debug removed and noted some things
Things I find wrong: the main screen needs to be larger. We should add a way back to the main menu after death. More enemies that come and go. some shoot some don't.  Rework the Score multipliers. instead of a instant multiplier maybe make it based on how many ships can be destroyed like for the next 10 the score is multiplied. Furthermore, I NEED TO add the new boss ideas as well as plug in the areas for sound and see where the sounds need to be added.  Also I removed some code that allowed the player to take damage during the invincibility timer.